finish profile component for now
